Refs #5858: _base/test_FocusManager.html "restore focus" button failing. This gets the code working on IE and FF, but not Safari...
The focus API tries to be clever, and (in this example)
saves not what is actually in focus (ie, the save button itself), but what was
in focus immediately before the save button was pressed.   That code fails on
Safari since the save button never gets focus.  Need to rethink or desupport this API
since I suspect the current code will fail on IE/FF if it's called
when nothing is in focus: it will save whatever used to be in focus, even if that
was 10 minutes earlier.  Will rethink for 2.0.

Patch partly from Nicola (CLA on file)